Adding side_ufuns() to all center ufuns
And adding MeanSMCenterUFun as an example of a ufun that is an arbitrary
function of the agreements that does not use individual ufuns for each
thread.

Even in this case, we cal also define a side-ufun for each thread (for
the center) that is simply the utility that the center would get if this
thread was the only one ending in agreement.