Adding make and run commands
The tournament command is divided into "make" and "run" subcommands.
Will later add "load"
For now, we run the command in make because loading tournaments from
yaml is not correct yet.

Also added center_multiplier and edge_multiplier in Tournament.run() to
control how much we weight center and edge runs. Default is to weigh
center runs n-edges times edge runs (center_multiplier is None).
This is not controllable from the CLI yet.