Adding expected-outcomes to center ufuns.
CenterUFun now keeps track of the expected-outcome for each thread.
This is initialized to None. It is possible to set the expected outcome
for any thread using CenterUFun.set_expected_outcome() or on a given
side directly using SideUFun.set_expected_outcome().

ANL2020Negotiator have now three parameters to control when to change
the expected-outcome of a thread.
1. update_expected_outcome_on_end: Sets it at the end of each thread.
   This defaults to True.
2. update_expected_outcome_after_offering: Sets it after offering. This
   defaults to False and is **not yet** implemented.
3. update_expected_outcome_on_receiving_offer: Sets it whenever we
   receive an offer before responding. This defaults to False and is
   **not yet** implemented.

Few supporting changes were needed. Please see the documentation of
ANL2020Negotiator:

- The context returned from self.negotiators and similar properties now
  have the `index` of the edge (besides the `ufun` and `center` which
  indicates whether or not this is a side of a center or an edge agent).