Always update expected outcomes
Before this commit, any center ufun with locally defined side-ufuns
(e.g. UtilityCombiningCenterUFuns) did not take into account the
expected-outcomes of other threads if the side-ufun is called directly
to evaluate an outcome.

We solved that by introducing SideUFunAdapter and encapsulating all
ufuns in such cases with these adapters. The adapters do take into
account the expected outcomes of other threads.

As an example of the value of that, we modified MaxCenterUFun so that
when the outcome is set in one thread, it updates the reserved value of
all other threads to be  at least the value of this outcome.