fixing expectation failure bug
For some reason that is not yet clear, deep copying the center ufun then
reconnecting sides does not always work. I avoided this copy completely.

A nagging issue: In my example run of makegraphs.py, the center gets an
agreement of 3 in the first run. It should offer 7 as the first offer in
the second thread but it offers 6. This has something to do with margins
in the PresortingInverseUtilityFunction. Will check that later. If we
can just assume that outcome-spaces are not huge, we can avoid using
this inverter completely and make our own.