[bugfix] Plotting upon mechanism completion
Originally, we plotted all graphs after **ALL** mechanisms are done
which means that expected-outcomes were set for all threads. This does
no map correctly to what side ufuns would be at the time of the
negotiation.

For this reason, we now plot figures and save them **immediately** upon
mechanism completion (which requires negmas 0.11.3 at least). This
should solve the bug. Note that this makes more sense with sequential
execution that with any other form of execution as the utility function
does change **during** the negotiation in all other types of execution
orders.