[bugfix] Avoiding a name-clash when loading multiple scenarios
All scenarios used center.py to implement the center evaluator when they
needed it. Because of the way python caches modules, this led sometimes
to an import error when trying to locate the new evaluator. We fixed
this by making sure every scenario that needs an evaluators implements
it in a different file